refactor: move protocols to separate file
