+++ src/components/Answer/protocols.ts
@@ -1,0 +1,11 @@
+export enum State {
+  Correct = 'Correct',
+  Wrong = 'Wrong',
+  Unavailable = 'Unavailable'
+}
+
+export type AnswerProps = {
+  text: string
+  answerIndex: number
+  correctAnswerIndex: number
+}
